cleanup: everything
- Remove py2
- Remove six
- Remove 2.8 duplicates
- Remove nosetests
- Remove deprecated build things
- Update github action stuff
