+++ .github/workflows/test.yml
@@ -10,13 +10,11 @@ jobs:
       - uses: actions/checkout@v2
       - uses: actions/setup-python@v2
         with:
-          python-version: '3.6'
+          python-version: '3.8'
       - name: Install requirements
         run: pip install flake8 pycodestyle pytest pytest-ckan pytest-cov
-      - name: Check syntax
-        run: flake8 . --count --select=E901,E999,F821,F822,F823 --show-source --statistics --exclude ckan,nose
       - name: Run flake8
-        run: flake8 . --count --max-line-length=127 --statistics --exclude ckan,nose
+        run: flake8 . --count --max-line-length=127 --statistics --exclude ckan
 
   lint-js:
     name: JS Lint and Build Metadata App
@@ -25,7 +23,7 @@ jobs:
       - uses: actions/checkout@v2
       - uses: actions/setup-python@v2
         with:
-          python-version: '3.6'
+          python-version: '3.8'
       - name: Install dependencies
         run: |
           yarn
@@ -41,7 +39,7 @@ jobs:
       - uses: actions/checkout@v2
       - uses: actions/setup-python@v2
         with:
-          python-version: '3.6'
+          python-version: '3.8'
       - name: Install dependencies
         run: |
           yarn
@@ -91,7 +89,6 @@ jobs:
         npm install --global yarn
         yarn install
     - name: Install requirements and Setup (py3)
-      if: ${{ matrix.ckan-version == '2.9' }}
       run: |
         pip install -r requirements.txt
         pip install -e .
