+++ ckanext/dcat_usmetadata/db_utils.py
@@ -1,7 +1,5 @@
 import ckan.model as model
 
-import six
-
 
 def get_parent_organizations(c):
     items = {}
@@ -20,10 +18,7 @@ def get_parent_organizations(c):
         ids = []
 
         for id in userGroupsIds:
-            if six.PY2:
-                ids.append(id.encode('ascii', 'ignore'))
-            else:
-                ids.append(id)
+            ids.append(id)
 
         # Ugly hack - If user has access to only one organization then SQL
         # query blows up because IN statement ends up with dangling comma at
