+++ ckanext/dcat_usmetadata/templates/new-metadata.html
