+++ ckanext/dcat_usmetadata/templates/snippets/add_dataset.html
