--- ckanext/dcat_usmetadata/templates/templates_2_8/organization/read.html
@@ -1,6 +1,0 @@
-{% ckan_extends %} {% block page_primary_action %} {% if h.check_access('package_create',
-{'owner_org': c.group_dict.id}) %}
-<a class="btn btn-primary" href="/dataset/new-metadata?group={{ c.group_dict.id }}"
-  >{{ _('Add Dataset') }}</a
->
-{% endif %} {{ super() }} {% resource 'dcat_usmetadata/custom.css' %} {% endblock %}
