--- ckanext/dcat_usmetadata/templates/templates_2_8/package/read_base.html
@@ -1,13 +1,0 @@
-{% ckan_extends %} {% block content_action %} {% if h.check_access('package_update', {'id':pkg.name
-}) %}
-<a class="btn btn-primary" href="/dataset/edit-new/{{ pkg.name }}">{{ _('Edit') }}</a>
-{% endif %} {% if h.check_access('package_delete', {'id':pkg.name}) %} {% set locale =
-h.dump_json({'content': _('Are you sure you want to delete this dataset?')}) %}
-<a
-  class="btn btn-danger"
-  href="{% url_for controller='package', action='delete', id=pkg.name %}"
-  data-module="confirm-action"
-  data-module-i18n="{{ locale }}"
-  >{% block delete_button_text %}{{ _('Delete') }}{% endblock %}</a
->
-{% endif %} {% endblock %}
