--- ckanext/dcat_usmetadata/templates/templates_new/base.html
@@ -1,2 +1,0 @@
-{% ckan_extends %} {% block styles %} {{ super() }} {% asset 'dcat_usmetadata/styles' %} {% endblock
-%}
