--- ckanext/dcat_usmetadata/templates/templates_new/new-metadata.html
@@ -1,11 +1,0 @@
-<div
-  id="root"
-  data-apiUrl="{{g.site_url}}/api/3/action/"
-  data-ownerOrg="{{request.args.get('group')}}"
-  data-datasetId="{{id or ''}}"
-></div>
-<link rel="stylesheet" href="/css/main.chunk.css" type"text/css"> <link rel="stylesheet"
-href="/css/2.chunk.css" type"text/css">
-<script src="/js/2.chunk.js"></script>
-<script src="/js/main.chunk.js"></script>
-<script src="/js/runtime-main.js"></script>
