--- ckanext/dcat_usmetadata/templates/templates_new/snippets/add_dataset.html
@@ -1,3 +1,0 @@
-{# Adds 'Add Dataset' button #}
-
-<a class="btn btn-primary" href="/dataset/new-metadata">{{ _('Add Dataset') }}</a>
