--- config/production.ini
@@ -1,213 +1,0 @@
-#
-# CKAN - Pylons configuration
-#
-# These are some of the configuration options available for your CKAN
-# instance. Check the documentation in 'doc/configuration.rst' or at the
-# following URL for a description of what they do and the full list of
-# available options:
-#
-# http://docs.ckan.org/en/latest/maintaining/configuration.html
-#
-# The %(here)s variable will be replaced with the parent directory of this file
-#
-
-[DEFAULT]
-
-# WARNING: *THIS SETTING MUST BE SET TO FALSE ON A PRODUCTION ENVIRONMENT*
-debug = false
-
-[server:main]
-use = egg:Paste#http
-host = 0.0.0.0
-port = 5000
-
-[app:main]
-use = egg:ckan
-full_stack = true
-cache_dir = /tmp/%(ckan.site_id)s/
-beaker.session.key = ckan
-
-# This is the secret token that the beaker library uses to hash the cookie sent
-# to the client. `paster make-config` generates a unique value for this each
-# time it generates a config file.
-beaker.session.secret = EqF8Z9z+c6lsEKHJzGJygghqz
-
-# `paster make-config` generates a unique value for this each time it generates
-# a config file.
-app_instance_uuid = 054d2b22-a3b9-4698-853d-63bbd151ecc1
-
-# repoze.who config
-who.config_file = %(here)s/who.ini
-who.log_level = warning
-who.log_file = %(cache_dir)s/who_log.ini
-# Session timeout (user logged out after period of inactivity, in seconds).
-# Inactive by default, so the session doesn't expire.
-# who.timeout = 86400
-
-## Database Settings
-sqlalchemy.url = postgresql://ckan:pass@db/ckan
-
-ckan.datastore.write_url = postgresql://datastore:pass@datastore/datastore
-ckan.datastore.read_url = postgresql://datastore_ro:pass@datastore/datastore
-
-# PostgreSQL' full-text search parameters
-ckan.datastore.default_fts_lang = english
-ckan.datastore.default_fts_index_method = gist
-
-## Site Settings
-
-ckan.site_url = http://127.0.0.1:5000
-#ckan.use_pylons_response_cleanup_middleware = true
-
-## Authorization Settings
-
-ckan.auth.anon_create_dataset = false
-ckan.auth.create_unowned_dataset = false
-ckan.auth.create_dataset_if_not_in_organization = false
-ckan.auth.user_create_groups = false
-ckan.auth.user_create_organizations = false
-ckan.auth.user_delete_groups = true
-ckan.auth.user_delete_organizations = true
-ckan.auth.create_user_via_api = false
-ckan.auth.create_user_via_web = true
-ckan.auth.roles_that_cascade_to_sub_groups = admin
-
-
-## Search Settings
-
-ckan.site_id = inventory
-solr_url = http://solr:8983/solr/inventory
-
-#ckan.simple_search = 1
-
-
-## CORS Settings
-
-# If cors.origin_allow_all is true, all origins are allowed.
-# If false, the cors.origin_whitelist is used.
-ckan.cors.origin_allow_all = true
-# cors.origin_whitelist is a space separated list of allowed domains.
-# ckan.cors.origin_whitelist = http://example1.com http://example2.com
-
-
-## Plugins Settings
-
-# Note: Add ``datastore`` to enable the CKAN DataStore
-#       Add ``datapusher`` to enable DataPusher
-#               Add ``resource_proxy`` to enable resorce proxying and get around the
-#               same origin policy
-ckan.plugins = usmetadata datajson stats text_view recline_view dcat_usmetadata
-
-
-# Define which views should be created by default
-# (plugins must be loaded in ckan.plugins)
-ckan.views.default_views = recline_view text_view image_view webpage_view recline_grid_view
-
-## Front-End Settings
-ckan.site_title = CKAN
-ckan.site_logo = /base/images/ckan-logo.png
-ckan.site_description =
-ckan.favicon = /images/icons/ckan.ico
-ckan.gravatar_default = identicon
-ckan.preview.direct = png jpg gif
-ckan.preview.loadable = html htm rdf+xml owl+xml xml n3 n-triples turtle plain atom csv tsv rss txt json
-ckan.display_timezone = server
-
-# package_hide_extras = for_search_index_only
-#package_edit_return_url = http://another.frontend/dataset/<NAME>
-#package_new_return_url = http://another.frontend/dataset/<NAME>
-#ckan.recaptcha.version = 1
-#ckan.recaptcha.publickey =
-#ckan.recaptcha.privatekey =
-#licenses_group_url = http://licenses.opendefinition.org/licenses/groups/ckan.json
-# ckan.template_footer_end =
-
-
-## Internationalisation Settings
-ckan.locale_default = en
-ckan.locale_order = en pt_BR ja it cs_CZ ca es fr el sv sr sr@latin no sk fi ru de pl nl bg ko_KR hu sa sl lv
-ckan.locales_offered =
-ckan.locales_filtered_out = en_GB
-
-## Feeds Settings
-
-ckan.feeds.authority_name =
-ckan.feeds.date =
-ckan.feeds.author_name =
-ckan.feeds.author_link =
-
-## Storage Settings
-
-ckan.storage_path = /var/lib/ckan
-#ckan.max_resource_size = 10
-#ckan.max_image_size = 2
-
-## Datapusher settings
-
-# Make sure you have set up the DataStore
-
-# ckan.datapusher.formats = csv xls xlsx tsv application/csv application/vnd.ms-excel application/vnd.openxmlformats-officedocument.spreadsheetml.sheet
-# ckan.datapusher.url = http://127.0.0.1:8800/
-
-# Resource Proxy settings
-# Preview size limit, default: 1MB
-#ckan.resource_proxy.max_file_size = 1048576
-# Size of chunks to read/write.
-#ckan.resource_proxy.chunk_size = 4096
-
-## Activity Streams Settings
-
-#ckan.activity_streams_enabled = true
-#ckan.activity_list_limit = 31
-#ckan.activity_streams_email_notifications = true
-#ckan.email_notifications_since = 2 days
-ckan.hide_activity_from_users = %(ckan.site_id)s
-
-# Data.json settings
-ckanext.datajson.inventory_links_enabled = True
-
-## Email settings
-
-#email_to = you@yourdomain.com
-#error_email_from = paste@localhost
-#smtp.server = localhost
-#smtp.starttls = False
-#smtp.user = your_username@gmail.com
-#smtp.password = your_password
-#smtp.mail_from =
-
-
-## Logging configuration
-[loggers]
-keys = root, ckan, ckanext
-
-[handlers]
-keys = console
-
-[formatters]
-keys = generic
-
-[logger_root]
-level = WARNING
-handlers = console
-
-[logger_ckan]
-level = INFO
-handlers = console
-qualname = ckan
-propagate = 0
-
-[logger_ckanext]
-level = DEBUG
-handlers = console
-qualname = ckanext
-propagate = 0
-
-[handler_console]
-class = StreamHandler
-args = (sys.stderr,)
-level = NOTSET
-formatter = generic
-
-[formatter_generic]
-format = %(asctime)s %(levelname)-5.5s [%(name)s] %(message)s
