--- requirements-py2.txt
@@ -1,32 +1,0 @@
--e git+https://github.com/GSA/ckanext-harvest.git@ckan-master#egg=ckanext-harvest
--e git+https://github.com/GSA/ckanext-datajson.git@main#egg=ckanext-datajson
--e git+https://github.com/GSA/ckanext-usmetadata.git@main#egg=ckanext-usmetadata
--e git+https://github.com/GSA/ckanext-googleanalyticsbasic.git@main#egg=ckanext-googleanalyticsbasic
--e git+https://github.com/GSA/ckanext-saml2.git@max#egg=ckanext-saml2
--e git+https://github.com/GSA/webob.git@ckan-patch#egg=webob
-
-
-repoze.lru==0.6
-repoze.who==2.0.0
-repoze.who-friendlyform==1.0.8
-
-# ckanext-saml2 dependencies
-M2Crypto==0.36.0
--e git+https://github.com/GSA/pysaml2.git@max#egg=pysaml2
-
-# ckanext-datajson dependencies
-pyyaml==5.4
-lepl
-jsonschema~=2.4.0
-rfc3987
-
-# ckanext-harvest dependencies
-pika
-redis
-requests
-pyOpenSSL
-ckantoolkit==0.0.3
-
-# not directly required, pinned by Snyk to avoid a vulnerability
-# or dev-requirements
-pip==20.3.3
