+++ requirements.txt
@@ -20,11 +20,9 @@ pika>=1.1.0
 pyOpenSSL==20.0.1
 # redis==2.10.6 # included in ckan core
 # requests>=2.11.1 # included in ckan core
-six>=1.12.0
 
 # not directly required, pinned by Snyk to avoid a vulnerability
 # or dev-requirements
 python-memcached==1.58
 pip==20.3.3
 flake8
-
