+++ setup.cfg
@@ -5,9 +5,6 @@ omit=ckanext/geodatagov/tests/*
 [flake8]
 max-line-length = 127
 
-[tool:pytest]
-norecursedirs=ckanext/geodatagov/tests/nose
-
 [extract_messages]
 keywords = translate isPlural
 add_comments = TRANSLATORS:
